Rename stateManager.ts to merkleStateManager.ts
