+++ packages/statemanager/src/index.ts
@@ -3,5 +3,5 @@ export * from './cache/index.js'
 export * from './rpcStateManager.js'
 export * from './simpleStateManager.js'
 export * from './statelessVerkleStateManager.js'
-export * from './stateManager.js'
+export * from './merkleStateManager.js'
 export * from './types.js'
